Replace nancy selfhost with owin
Let OWIN handle CORS and request logging
