+++ src/compute.geometry/Program.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
+using Microsoft.Owin.Hosting;
 using Nancy;
-using Nancy.Bootstrapper;
-using Nancy.Conventions;
-using Nancy.Extensions;
-using Nancy.Hosting.Self;
 using Nancy.Routing;
-using Nancy.TinyIoc;
 using Serilog;
 using Topshelf;
 
@@ -23,37 +20,47 @@ namespace compute.geometry
             RhinoInside.Resolver.Initialize();
 #if DEBUG
             string rhinoSystemDir = @"C:\dev\github\mcneel\rhino\src4\bin\Debug";
-            if (System.IO.Directory.Exists(rhinoSystemDir))
+            if (File.Exists(rhinoSystemDir + "\\Rhino.exe"))
                 RhinoInside.Resolver.RhinoSystemDirectory = rhinoSystemDir;
 #endif
 
             Logging.Init();
             LogVersions();
 
-            // use a different backend port for debugging, so we don't have to reserve/unreserve ports when testing in Release
-#if DEBUG
-            int defaultBackendPort = 8082;
-#else
-            int defaultBackendPort = 8081;
-#endif
-            int backendPort = Env.GetEnvironmentInt("COMPUTE_BACKEND_PORT", defaultBackendPort);
+            int port = Env.GetEnvironmentInt("COMPUTE_BACKEND_PORT", 8081);
 
-            Topshelf.HostFactory.Run(x =>
+            //            Topshelf.HostFactory.Run(x =>
+            //            {
+            //                x.UseSerilog();
+            //                x.ApplyCommandLine();
+            //                x.SetStartTimeout(new TimeSpan(0, 1, 0));
+            //                x.Service<NancySelfHost>(s =>
+            //                  {
+            //                      s.ConstructUsing(name => new NancySelfHost());
+            //                      s.WhenStarted(tc => tc.Start(port));
+            //                      s.WhenStopped(tc => tc.Stop());
+            //                  });
+            //                x.RunAsPrompt();
+            //                //x.RunAsLocalService();
+            //                x.SetDisplayName("compute.geometry");
+            //                x.SetServiceName("compute.geometry");
+            //            });
+
+            var url = $"http://localhost:{port}";
+
+            StartOptions options = new StartOptions(url);
+
+            // disable built-in owin tracing by using a null traceoutput
+            options.Settings.Add(
+                typeof(Microsoft.Owin.Hosting.Tracing.ITraceOutputFactory).FullName,
+                typeof(NullTraceOutputFactory).AssemblyQualifiedName);
+
+            //using (WebApp.Start<Startup>(url))
+            using (WebApp.Start<Startup>(options))
             {
-                x.UseSerilog();
-                x.ApplyCommandLine();
-                x.SetStartTimeout(new TimeSpan(0, 1, 0));
-                x.Service<NancySelfHost>(s =>
-                  {
-                      s.ConstructUsing(name => new NancySelfHost());
-                      s.WhenStarted(tc => tc.Start(backendPort));
-                      s.WhenStopped(tc => tc.Stop());
-                  });
-                x.RunAsPrompt();
-                //x.RunAsLocalService();
-                x.SetDisplayName("compute.geometry");
-                x.SetServiceName("compute.geometry");
-            });
+                Log.Information("Running on {Url}", url);
+                Console.ReadLine();
+            }
 
             if (RhinoCore != null)
                 RhinoCore.Dispose();
@@ -72,121 +79,11 @@ namespace compute.geometry
         }
     }
 
-    public class NancySelfHost
+    public class NullTraceOutputFactory : Microsoft.Owin.Hosting.Tracing.ITraceOutputFactory
     {
-        private NancyHost _nancyHost;
-        private System.Diagnostics.Process _backendProcess = null;
-
-        public void Start(int http_port)
+        public TextWriter Create(string outputFile)
         {
-            Log.Information("Launching RhinoCore library as {User}", Environment.UserName);
-            Program.RhinoCore = new Rhino.Runtime.InProcess.RhinoCore(null, Rhino.Runtime.InProcess.WindowStyle.NoWindow);
-            var config = new HostConfiguration();
-#if DEBUG
-            config.RewriteLocalhost = false;  // Don't require URL registration since geometry service always runs on localhost
-#endif
-            var listenUriList = new List<Uri>();
-
-            if (http_port > 0)
-                listenUriList.Add(new Uri($"http://localhost:{http_port}"));
-
-            if (listenUriList.Count > 0)
-                _nancyHost = new NancyHost(config, listenUriList.ToArray());
-            else
-                Log.Error("Neither COMPUTE_HTTP_PORT nor COMPUTE_HTTPS_PORT are set. Not listening!");
-            try
-            {
-                _nancyHost.Start();
-                foreach (var uri in listenUriList)
-                    Log.Information("compute.geometry running on {Uri}", uri.OriginalString);
-            }
-            catch (AutomaticUrlReservationCreationFailureException)
-            {
-                Log.Error(GetAutomaticUrlReservationCreationFailureExceptionMessage(listenUriList));
-                Environment.Exit(1);
-            }
-        }
-
-        // TODO: move this somewhere else
-        string GetAutomaticUrlReservationCreationFailureExceptionMessage(List<Uri> listenUriList)
-        {
-            var msg = new StringBuilder();
-            msg.AppendLine("Url not reserved. From an elevated command promt, run:");
-            msg.AppendLine();
-            foreach (var uri in listenUriList)
-                msg.AppendLine($"netsh http add urlacl url=\"{uri.Scheme}://+:{uri.Port}/\" user=\"Everyone\"");
-            return msg.ToString();
-        }
-
-        public void Stop()
-        {
-            if (_backendProcess != null)
-                _backendProcess.Kill();
-            _nancyHost.Stop();
-        }
-    }
-
-    public class Bootstrapper : Nancy.DefaultNancyBootstrapper
-    {
-        private byte[] _favicon;
-
-        protected override void ApplicationStartup(TinyIoCContainer container, IPipelines pipelines)
-        {
-            Log.Debug("ApplicationStartup");
-
-            // Load GH at startup so it can get initialized on the main thread
-            var pluginObject = Rhino.RhinoApp.GetPlugInObject("Grasshopper");
-            var runheadless = pluginObject?.GetType().GetMethod("RunHeadless");
-            if (runheadless != null)
-                runheadless.Invoke(pluginObject, null);
-
-            var loadComputePlugsin = typeof(Rhino.PlugIns.PlugIn).GetMethod("LoadComputeExtensionPlugins");
-            if (loadComputePlugsin != null)
-                loadComputePlugsin.Invoke(null, null);
-
-            Nancy.StaticConfiguration.DisableErrorTraces = false;
-            pipelines.OnError += (ctx, ex) => LogError(ctx, ex);
-            pipelines.AfterRequest += AddCORSSupport;
-            ApiKey.Initialize(pipelines);
-
-
-            Rhino.Runtime.HostUtils.RegisterComputeEndpoint("grasshopper", typeof(Endpoints.GrasshopperEndpoint));
-
-            base.ApplicationStartup(container, pipelines);
-        }
-
-        private static void AddCORSSupport(NancyContext context)
-        {
-            context.Response.Headers.Add("Access-Control-Allow-Origin", "*");
-            context.Response.Headers.Add("Access-Control-Allow-Methods", "OPTIONS,POST,GET,HEAD");
-            context.Response.Headers.Add("Access-Control-Allow-Headers", "Authorization,Origin,Accept,Content-Type,User-Agent,Access-Control-Allow-Headers,Access-Control-Request-Method,Access-Control-Request-Headers");
-        }
-        protected override void ConfigureConventions(NancyConventions nancyConventions)
-        {
-            base.ConfigureConventions(nancyConventions);
-            nancyConventions.StaticContentsConventions.Add(StaticContentConventionBuilder.AddDirectory("docs"));
-        }
-
-        protected override byte[] FavIcon
-        {
-            get { return _favicon ?? (_favicon = LoadFavIcon()); }
-        }
-
-        private byte[] LoadFavIcon()
-        {
-            using (var resourceStream = GetType().Assembly.GetManifestResourceStream("compute.geometry.favicon.ico"))
-            {
-                var memoryStream = new System.IO.MemoryStream();
-                resourceStream.CopyTo(memoryStream);
-                return memoryStream.GetBuffer();
-            }
-        }
-
-        private static dynamic LogError(NancyContext ctx, Exception ex)
-        {
-            string id = ctx.Request.Headers["X-Compute-Id"].FirstOrDefault();
-            Log.Error(ex, "An exception occured while processing request \"{RequestId}\"", id);
-            return null;
+            return StreamWriter.Null;
         }
     }
