+++ src/compute.geometry/compute.geometry.csproj
@@ -49,6 +49,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ApiKey.cs" />
+    <Compile Include="Bootstrapper.cs" />
     <Compile Include="DataCache.cs" />
     <Compile Include="Endpoints\GrasshopperEndpoint.cs" />
     <Compile Include="Environment.cs" />
@@ -60,6 +61,7 @@
     <Compile Include="Logging.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Startup.cs" />
     <EmbeddedResource Include="RhinoCompute.cs" />
     <Compile Include="ResthopperEndpoints.cs" />
   </ItemGroup>
@@ -76,13 +78,22 @@
     <PackageReference Include="AWSSDK.S3">
       <Version>3.3.24</Version>
     </PackageReference>
+    <PackageReference Include="Microsoft.Owin.Cors">
+      <Version>4.0.1</Version>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Owin.Host.HttpListener">
+      <Version>4.0.1</Version>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Owin.Hosting">
+      <Version>4.0.1</Version>
+    </PackageReference>
     <PackageReference Include="Nancy">
       <Version>1.4.4</Version>
     </PackageReference>
     <PackageReference Include="Nancy.Gzip">
       <Version>0.1.0</Version>
     </PackageReference>
-    <PackageReference Include="Nancy.Hosting.Self">
+    <PackageReference Include="Nancy.Owin">
       <Version>1.4.1</Version>
     </PackageReference>
     <PackageReference Include="Newtonsoft.Json">
